requirements.txt
streamlit
pandas
python-dotenv
google-generativeai
pypdf2
